Adding a clean up tool to process lots of files
This tool is a wrapper to find and sed. Repeated clean up tasks as
replacing tabs with spaces or removing trailing white space can be done
in a controlled manner over lots of files. No need to remember or
meditate for the same regexp again. You can store more expressions in
a custom config file.