variables in reclass and node live status
Introduce variables in the reclass hierarchy that can be filled with
real values from the local config file. Further we now have a way to
show the installation status of a node by connecting via ssh an reading
the distribution description via lsb_release (or complain about a
missing component..).