custom merge mode
This is a quite unusual use case, I have to admit.
It adds a custom merge mode that lets the user specify custom merge rule
in from of a reclass dict for any config file to be merged into a file
hierarchy structure that may than serve for custom configuration
engines, e.g. as in our case, a git repository that gets pushed to a
site under the control of a third party computation centre. From there
it will be used by salt states as input files, but we do not have direct
access to the states (or reclass, or pillars, or grains) themselves.