test protocols
Explicitly test for protocol support and compare the result to the list
of configured protocol names considered safe, ok, or unsafe. The list
must be kept up to date by the user. TODO: The red dot in the ciphers
test was removed, as for a preparation to unify the behaviour, we should
be able to keep a separate record for safe/unsafe ciphers too, currently
it only shows supported/unsupported.