Adding backupninja handlers
Adding our custom backupninja handlers that are running unchanged
for years now.

The ldap handler was necessary as we started to use the new config
style with ldap.

The collect handler is basically an rsync based synchronisation tool.
It was written at a time there was no 'format=mirror' in upstream. It
also saves the permissions to a separte file. We push the backups to
a central backup host, from where the actual backupserver pulls the
backups afterwards. Thus combining the advantages of both push and
pull backup. Both parts are done with the collect handler.