Clone and init knowledgebase dirs
Use the newly created functionality to both clone and init new
directories. While in init mode, the directory is probably used for
connecting ansible/debops/.. to the reclass knowledge base, the clone
mode can be used to separate or duplicate the original reclass
hierarchy, a separate nodeify config file is also put in place.