Reorganizing almost all variables in reclass
Ansible is quite restrictive in what to accept as variables. We need to
change the recommended "." to "__" in reclass to comply with all our
config and ansible. As we already had "__" for the bash restriction
for not using "-" in variables, there was another conflict now to
solve, now use "-" in reclass and change it explicitly for bash to
"_".