Pass local dirs on to ansible
In reclass we defined variables only used in this script until now.
These variables are used to be replaced by working directories of all
sorts of repositories on the local machine. As we now want to make use
this information also in the playbooks, we need to pass them on as
ansible extra vars on the commandline.